Fix double publishing
Ref https://github.com/sbt/sbt-release/issues/219

I don't think double publishing is the problem of sbt-release.

Part of the reason double publishing is happening is due to `crossScalaVersions := Seq(scala212, scala211)` set in artificial subprojects like `root` and `shaded`, and that these subprojects aggregate shaded Java subprojects. This means that when `+` traversal is on 2.11.x, it will propagate `publishSigned` to them, resulting in double publishing.

This adds `crossScalaVersions := Seq(scala212)` to `disablePublishing` so `root` and `shaded` do not cross build.

In addition, to avoid command propagation `aggregate(shaded)` was removed from `play-ahc-ws-standalone`.