Fix double publishing (#243)
* SBT version 0.13.15 --> 1.0.4 (#210)

* SBT version 0.13.15 --> 1.0.4

This change upgrades all SBT plugins -- including [sbt-scalariform]

[sbt-scalariform] configuration has changed.

* use 'publishSigned'

* Travis:  use 'testOnly' instead of 'test-only'

* sbt version 1.0.4 -> 1.1.0 (#223)

* Fix release process (#236)

* Avoid cross build for shaded projects (#237)

Related to sbt/sbt#3431

* sbt 1.1.1

* Fix double publishing

Ref https://github.com/sbt/sbt-release/issues/219

I don't think double publishing is the problem of sbt-release.

Part of the reason double publishing is happening is due to `crossScalaVersions := Seq(scala212, scala211)` set in artificial subprojects like `root` and `shaded`, and that these subprojects aggregate shaded Java subprojects. This means that when `+` traversal is on 2.11.x, it will propagate `publishSigned` to them, resulting in double publishing.

This adds `crossScalaVersions := Seq(scala212)` to `disablePublishing` so `root` and `shaded` do not cross build.

In addition, to avoid command propagation `aggregate(shaded)` was removed from `play-ahc-ws-standalone`.