Use a browsable URL for scm.url
This changes the scm.url to a browsable URL as recommended by the [POM Reference](https://maven.apache.org/pom.html#SCM). Scala Steward uses this URL to link to the updated dependency (as can be seen [here](https://github.com/mfirry/scala-http-clients/pull/76)) and a `git@` URL is not suitable for this purpose.